[REFACTOR] 쿠폰 이벤트 및 사용자 관련 로직 리팩토링 (#40)

+++ src/main/kotlin/kr/hhplus/be/server/application/couponevent/dto/CouponEventDTO.kt
@@ -4,14 +4,14 @@ data class CouponEventDTO(
     val id: String,
     val benefitMethod: String,
     val benefitAmount: String,
-    val totalIssueAmount: Int,
-    val leftIssueAmount: Int
+    val totalIssueAmount: Long,
+    val leftIssueAmount: Long
 )
 
 data class CreateCouponEventCriteria(
     val benefitMethod: String,
     val benefitAmount: String,
-    val totalIssueAmount: Int
+    val totalIssueAmount: Long,
 )
 
 data class IssueCouponCriteria(
