+++ src/main/kotlin/kr/hhplus/be/server/controller/couponevent/request/CreateCouponEventRequest.kt
@@ -3,6 +3,6 @@ package kr.hhplus.be.server.controller.couponevent.request
 data class CreateCouponEventRequest(
     val benefitMethod: String,
     val benefitAmount: String,
-    val totalIssueAmount: Int
+    val totalIssueAmount: Long
 )
